% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dycoef.R
\name{dycoef}
\alias{dycoef}
\title{Sequential Display of Coefficient Estimates}
\usage{
dycoef(
  data,
  y,
  x,
  covar = NULL,
  time_id,
  years = seq(min(data[[time_id]]), max(data[[time_id]]), 1),
  linetype = 6,
  size = 3
)
}
\arguments{
\item{data}{a data.frame object}

\item{y}{the outcome variable of your interest}

\item{x}{the explaining variable of your interest, should be available across several time periods}

\item{covar}{control variables to be included}

\item{time_id}{ID for each time point}

\item{years}{a sequence of periods you want to take into account; all years in your data by default.}
}
\value{
This functions return the following three objects:
A. "lastreg": regression results of you model in the subsample of last time period
B. "results": A data frame reporting coefficients and 95\% CIs for all regressions.
C. "plot": A dynamic plot of x coefficients across periods using ggplot2().
}
\description{
Running one regresison model over multiple subsamples by time periods, get the resuls, and plot coefficients over time.
}
\examples{

# load in your data
anes <- read.dta13("data/ANES cleaned.dta")

# use the function
anes.dynamic <- dycoef(
  data = anes, dv = "pid7", x = 'ft_black', time_id = "year",
  years = seq(1960,2020,4),
  covar = ('ft_white','as.factor(race)')
)

# see the organized table for coefficients and 95\% CIs
anes.dynamic$results

# see the dynamic plot of DiD results
anes.dynamic$plot

}
