% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dydid.R
\name{dydid}
\alias{dydid}
\title{dydid}
\usage{
dydid(
  data = df,
  dv = "y",
  tpoint = "reform_timepoint",
  step = 1,
  eid = "state_id",
  tid = "year_id",
  span = c(-5, 5),
  covar = NULL
)
}
\arguments{
\item{data}{a data set object}

\item{dv}{the outcome variable of your interest}

\item{tpoint}{variable in your data that indicates the time when the treatment takes place}

\item{step}{the unit by which the time ID increments, e.g., 2 for every two years}

\item{eid}{ID for each individual entity}

\item{tid}{ID for each time point}

\item{span}{the number of periods before and after the treatment you want to take into consideration.}

\item{covar}{controls variables to be included}
}
\value{
This functions return the following three objects:
A. "regs": Original results of DiD regressions using felm().
B. "results": A data frame reporting coefficients and 95% CIs for treatment-time indicators.
C. "plot": A dynamic plot of DiD coefficients using ggplot2().
}
\description{
Set specifications for dynamic Difference-in-Differences analysis, report coefficients from regressions, and draw coefficients in a dynamic way.
}
\examples{

# load in your data
df <- read.dta13("data/big bad banks.dta")

# use the function
bbb.did <- dydid(
  data = df, dv = "y", tpoint = "reform_timepoint", step = 1,
  eid = "state_id", tid = "year_id",
  span = c(-7, 7), covar = NULL
)

# see the origigal regression results
bbb.did$regs

# see the organized table for coefficients and 95\% CIs
bbb.did$results

# see the dynamic plot of DiD results
bbb.did$plot

}
