% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dycoef.R
\name{dycoef}
\alias{dycoef}
\title{dycoef}
\usage{
dycoef(
  data,
  dv = "pid7",
  x = "ft_black",
  covar = NULL,
  tid,
  years = seq(min(anes[[tid]]), max(anes[[tid]]), 1)
)
}
\arguments{
\item{data}{a data set object}

\item{dv}{the outcome variable of your interest}

\item{x}{the explaining variable of your interest, should be available across several time periods}

\item{covar}{controls variables to be included}

\item{tid}{ID for each time point}

\item{years}{a sequence of perids you want to take into account; all years in your data by default.}
}
\value{
This functions return the following three objects:
A. "lastreg": regression results of you model in the subsample of last time period
B. "results": A data frame reporting coefficients and 95% CIs for all regressions.
C. "plot": A dynamic plot of x coefficients across periods using ggplot2().
}
\description{
Running one regresison model over multiple subsamples by time periods, get the resuls, and plot coefficients over time.
}
\examples{

# load in your data
anes <- read.dta13("data/ANES cleaned.dta")

# use the function
anes.dynamic <- dycoef(
  data = anes, dv = "pid7", x = 'ft_black', tid = "year",
  years = seq(1960,2020,4) ,
  covar = ('ft_white', 'as.factor(race)')
)

# see the organized table for coefficients and 95\% CIs
anes.dynamic$results

# see the dynamic plot of DiD results
anes.dynamic$plot

}
